requirements.txt
amqp==5.1.1
argon2-cffi==20.1.0
arrow==1.2.3
asgiref==3.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==3.6.4.0
binaryornot==0.4.4
boto3==1.17.97
botocore==1.20.112
celery==5.1.2
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
chardet==5.2.0
charset-normalizer==3.2.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cookiecutter==1.7.3
crispy-tailwind==0.5.0
cryptography==41.0.3
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==3.2.2
django-allauth==0.56.1
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-clearcache==1.2.1
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-js-asset==2.1.0
django-redis==5.0.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-taggit==1.5.1
django-tailwind==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
env==0.1.0
filelock==3.12.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
kombu==5.3.2
Markdown==3.3.4
MarkupSafe==2.1.3
mutagen==1.45.1
numpy==1.25.2
oauthlib==3.2.2
pandas==2.1.0
Pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
poyo==0.5.0
prompt-toolkit==3.0.39
psycopg2==2.8.6
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.1
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2020.5
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.4.2
service-identity==23.1.0
setuptools==65.5.0
six==1.16.0
sqlparse==0.3.1
stripe==2.58.0
text-unidecode==1.3
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.4
wcwidth==0.2.6
whitenoise==5.2.0
zope.interface==6.0
gunicorn 
whitenoise== 5.2.0